+++ lib/src/core/keys/api_keys.dart
@@ -6,9 +6,7 @@ import 'package:uni_pay/uni_pay.dart';
 class ApiKeys {
   ///* Keys -------------------------------
   static String get tamaraBaseUrl =>
-      UniPayControllers.uniPayData.environment.isProduction
-          ? tamaraProductionBaseUrl
-          : tamaraBaseUrlForDev;
+      UniPayControllers.uniPayData.environment.tamaraBaseUrl;
 
   static String tamaraProductionBaseUrl = "https://api.tamara.co";
   static String tamaraBaseUrlForDev = "https://api-sandbox.tamara.co";
@@ -20,25 +18,26 @@ class ApiKeys {
   static const String failedUrl = "$uniBaseURl/failed";
   static const String cancelUrl = "$uniBaseURl/cancelled";
   static String approved = "approved";
-  static String get tamaraCheckoutUrl =>
-      "https://api-sandbox.tamara.co/checkout";
+  static String get tamaraCheckoutUrl => "$tamaraBaseUrl/checkout";
 
   static Map<String, String> get tamaraHeaders => UniPayControllers
       .uniPayData.credentials.tamaraCredential!.token.tamaraHeaders;
 
   ///* Process payment --------------------------------
   static PaymentConfig get moyasarPaymentConfig => PaymentConfig(
-      publishableApiKey: UniPayControllers
-          .uniPayData.credentials.moyasarCredential!.publishableKey,
-      amount: UniPayControllers
-          .uniPayData.orderInfo.transactionAmount.totalAmount.amountInHalala,
-      description: UniPayControllers.uniPayData.orderInfo.description,
-      applePay: ApplePayConfig(
-        label: UniPayControllers.uniPayData.appName,
-        merchantId: UniPayControllers
-                .uniPayData.credentials.applePayMerchantIdentifier ??
-            "",
-      ));
+        publishableApiKey: UniPayControllers
+            .uniPayData.credentials.moyasarCredential!.publishableKey,
+        amount: UniPayControllers
+            .uniPayData.orderInfo.transactionAmount.totalAmount.amountInHalala,
+        description: UniPayControllers.uniPayData.orderInfo.description,
+        applePay: ApplePayConfig(
+          label: UniPayControllers.uniPayData.appName,
+          merchantId: UniPayControllers
+                  .uniPayData.credentials.applePayMerchantIdentifier ??
+              "",
+        ),
+        metadata: {"orderId": UniPayControllers.uniPayData.orderInfo.orderId},
+      );
 
   ///* Web view options
   static InAppWebViewGroupOptions webViewGroupOptions =
@@ -49,4 +48,8 @@ class ApiKeys {
     ios: IOSInAppWebViewOptions(
         applePayAPIEnabled: true, allowsInlineMediaPlayback: true),
   );
+
+  ///* Moyasar
+  static String moyasarBaseUrl = "https://api.moyasar.com/v1";
+  static String moyasarPaymentsUrl = "$moyasarBaseUrl/payments";
 }