+++ lib/src/core/models/meta_data.dart
@@ -1,0 +1,9 @@
+class UniPayMetaData {
+  late String key;
+  late String value;
+
+  UniPayMetaData({
+    required this.key,
+    required this.value,
+  });
+}