+++ lib/src/core/services/uni_pay_services.dart
@@ -1,6 +1,6 @@
-import 'package:uni_pay/src/modules/tamara/core/models/capture_order.dart';
-import 'package:uni_pay/src/modules/tamara/core/models/capture_response.dart';
-import 'package:uni_pay/src/modules/tamara/core/services/uni_tamara.dart';
+import 'package:uni_pay/uni_pay.dart';
+
+import '../../modules/moyasar/core/services/uni_moyasar.dart';
 
 class UniPayServices {
   UniPayServices._();
@@ -9,5 +9,20 @@ class UniPayServices {
   static Future<TamaraCaptureOrderResponse> captureTamaraOrder(
       {required TamaraCaptureOrder tamaraCaptureOrder}) async {
     return await UniTamara.captureOrder(tamaraCaptureOrder);
+  }
+
+  ///* Get the payment by meta-data
+  static Future<UniPayResponse> getPaymentDetailsByMetaDataOrderId({
+    required MoyasarCredential credential,
+    required String orderId,
+  }) async {
+    return UniPayMoyasarGateway.searchPaymentByMetaData(
+        credential: credential, orderId: orderId);
+  }
+
+  ///* Get tamara order transaction information
+  static Future getTamaraPaymentDetails(
+      {required TamaraApiData tamaraApiData}) async {
+    return await UniTamara.getTransactionInfo(tamaraApiData);
   }
 }