+++ lib/src/modules/moyasar/core/services/uni_moyasar.dart
@@ -1,7 +1,14 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:moyasar/moyasar.dart';
 import 'package:uni_pay/src/constant/uni_text.dart';
+import 'package:uni_pay/src/utils/extension.dart';
 import 'package:uni_pay/uni_pay.dart';
+import 'package:http/http.dart' as http_client;
+
+import '../../../../core/keys/api_keys.dart';
+import '../../../../utils/utils.dart';
 
 class UniPayMoyasarGateway {
   UniPayMoyasarGateway._();
@@ -15,14 +22,38 @@ class UniPayMoyasarGateway {
         uniLog(result.status);
         if (result.status == PaymentStatus.paid) {
           uniPayResponse.transactionId = result.id;
-          uniPayResponse.amount = result.amount;
-          uniPayResponse.fee = result.fee;
+          uniPayResponse.amount = result.amount.halalaToAmount;
+          uniPayResponse.fee = result.fee.halalaToAmount;
           uniPayResponse.status = UniPayStatus.success;
           uniPayResponse.amountFormatted = result.amountFormat;
           uniPayResponse.ip = result.ip;
           uniPayResponse.createdAt = result.createdAt;
           uniPayResponse.invoiceId = result.invoiceId ?? "N/A";
           uniPayResponse.description = result.description ?? "N/A";
+          UniPayTransactionDetails transactionDetails =
+              UniPayTransactionDetails();
+          //* Card pay
+          final source = result.source;
+          if (source is CardPaymentResponseSource) {
+            transactionDetails.type = UniPayPaymentMethods.card;
+            transactionDetails.name = source.name;
+            transactionDetails.cardType = source.company.name.cardType;
+            transactionDetails.cardNumber = source.number;
+            transactionDetails.message = source.message ?? "";
+            transactionDetails.gatewayId = source.gatewayId;
+            transactionDetails.referenceNumber = source.referenceNumber ?? "";
+          }
+          //* Apple pay
+          else if (source is ApplePayPaymentResponseSource) {
+            transactionDetails.type = UniPayPaymentMethods.applepay;
+            transactionDetails.name = UniPayPaymentMethods.applepay.name;
+            transactionDetails.cardType = UniPayCardType.mada;
+            transactionDetails.cardNumber = source.number;
+            transactionDetails.message = source.message ?? "";
+            transactionDetails.gatewayId = source.gatewayId;
+            transactionDetails.referenceNumber = source.referenceNumber ?? "";
+          }
+          uniPayResponse.transactionDetails = transactionDetails;
         } else {
           uniPayResponse.status = UniPayStatus.failed;
           uniPayResponse.errorMessage = UniPayText.paymentFailed;
@@ -36,5 +67,37 @@ class UniPayMoyasarGateway {
       UniPayControllers.handlePaymentsResponseAndCallback(context,
           response: uniPayResponse);
     }
+  }
+
+  ///* Get the payment by meta-data
+  static Future<UniPayResponse> searchPaymentByMetaData({
+    required MoyasarCredential credential,
+    required String orderId,
+  }) async {
+    UniPayResponse uniPayResponse = UniPayResponse(
+      status: UniPayStatus.failed,
+      errorMessage: UniPayText.noTransactionFound,
+    );
+    try {
+      Uri paymentVerifyUrl = Uri.parse(
+          "${ApiKeys.moyasarPaymentsUrl}/?metadata[orderId]=$orderId");
+
+      Map<String, String> headerData = {
+        "authorization": credential.moyasarAuthKey
+      };
+      http_client.Response response =
+          await http_client.get(paymentVerifyUrl, headers: headerData);
+      Map<String, dynamic> responseData = json.decode(response.body) ?? {};
+      if (responseData["payments"] != null &&
+          responseData["payments"].isNotEmpty) {
+        uniPayResponse = UniPayResponse.fromMap(responseData["payments"].first);
+      } else {
+        uniPayResponse.errorMessage =
+            responseData["message"] ?? UniPayText.noTransactionFound;
+      }
+    } catch (e) {
+      uniPayResponse.errorMessage = e.toString();
+    }
+    return uniPayResponse;
   }
 }