+++ lib/src/modules/tamara/core/services/uni_tamara.dart
@@ -74,13 +74,14 @@ class UniTamara {
       merchantUrl: uniPayData.credentials.tamaraCredential!.merchantUrl,
     );
 
-    TamaraCheckoutData checkout = await UniTamara.tamaraCheckout(tamaraData);
+    TamaraCheckoutData checkout =
+        await UniTamara.tamaraCheckout(tamaraData, uniPayData);
     return checkout;
   }
 
   ///* Call tamara checkout
   static Future<TamaraCheckoutData> tamaraCheckout(
-      TamaraData tamaraData) async {
+      TamaraData tamaraData, UniPayData uniPayData) async {
     TamaraCheckoutData checkout = TamaraCheckoutData();
     try {
       String data = jsonEncode(tamaraData.toJson());
@@ -176,26 +177,55 @@ class UniTamara {
     TamaraCaptureOrderResponse captureOrderResponse =
         TamaraCaptureOrderResponse();
 
-    String bodyData = json.encode(tamaraCaptureOrder.toJson());
-    http_client.Response response = await http_client.post(
-      Uri.parse(tamaraCaptureOrder.environment.tamaraCapturePayment),
-      headers: tamaraCaptureOrder.tamaraToken.tamaraHeaders,
-      body: bodyData,
-    );
-    final data = json.decode(response.body);
-    if (response.statusCode.isSuccess) {
-      String orderId = data["order_id"] ?? "";
-      bool isSuccess =
-          orderId.isNotEmpty && orderId == tamaraCaptureOrder.orderId;
+    try {
+      String bodyData = json.encode(tamaraCaptureOrder.toJson());
+      http_client.Response response = await http_client.post(
+        Uri.parse(tamaraCaptureOrder.environment.tamaraCapturePayment),
+        headers: tamaraCaptureOrder.tamaraToken.tamaraHeaders,
+        body: bodyData,
+      );
+      final data = json.decode(response.body);
+      if (response.statusCode.isSuccess) {
+        String orderId = data["order_id"] ?? "";
+        bool isSuccess =
+            orderId.isNotEmpty && orderId == tamaraCaptureOrder.orderId;
 
-      captureOrderResponse =
-          TamaraCaptureOrderResponse.fromJson(data, isSuccess);
-    } else {
-      captureOrderResponse.message = data["message"] ?? "";
+        captureOrderResponse =
+            TamaraCaptureOrderResponse.fromJson(data, isSuccess);
+      } else {
+        captureOrderResponse.message = data["message"] ?? "";
+      }
+    } catch (e) {
+      captureOrderResponse.message = e.toString();
+    }
+    uniPrint("Capture called ${captureOrderResponse.toJson()}");
+    return captureOrderResponse;
+  }
+
+  ///* Get order information
+  static Future<UniPayResponse> getTransactionInfo(
+      TamaraApiData tamaraApiData) async {
+    UniPayResponse uniPayResponse =
+        UniPayResponse(status: UniPayStatus.notFound);
+
+    try {
+      http_client.Response response = await http_client.get(
+        Uri.parse(tamaraApiData.environment
+            .tamaraTransactionApi(tamaraApiData.orderId)),
+        headers: tamaraApiData.tamaraToken.tamaraHeaders,
+      );
+      final data = json.decode(response.body);
+      uniLog("Saif got this data -------->$data");
+      if (response.statusCode.isSuccess) {
+        uniPayResponse = UniPayResponse.fromTamara(data, tamaraApiData.orderId);
+      } else {
+        uniPayResponse.errorMessage = data["message"] ?? "No order found!";
+      }
+    } catch (e) {
+      uniPayResponse.errorMessage = e.toString();
     }
 
-    uniPrint(
-        "Capture called ${captureOrderResponse.toJson()}---> ${response.body}");
-    return captureOrderResponse;
+    uniLog("Get transaction called ${uniPayResponse.toMap()}");
+    return uniPayResponse;
   }
 }