+++ lib/src/utils/utils.dart
@@ -1,5 +1,3 @@
-import 'dart:developer';
-
 import 'package:flutter/foundation.dart';
 
 class Utils {
@@ -9,13 +7,13 @@ class Utils {
 ///* Print shortcuts `print()`
 void uniPrint(dynamic data) {
   if (kDebugMode) {
-    print(data);
+    // print(data);
   }
 }
 
 ///* Print in log shortcuts `log()`
 void uniLog(dynamic data) {
   if (kDebugMode) {
-    log(data.toString(), time: DateTime.now());
+    // log(data.toString(), time: DateTime.now());
   }
 }