+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.1.3
+
+- Added method to check tamara transaction information by ref id and fixes card type bugs
+
+## 1.1.2
+
+- Added card type for transaction(mada, visa, mastercard, and amex)
+
+## 1.1.1
+
+- Added card type for transaction(mada, visa, mastercard, and amex) also you can now fetch any payment info by orderId you provided
+
 ## 1.0.3
 
 - Added authorization and capture order for Tamara